+++ utils-test-app/utils_test_app/tests/test_esi.py
@@ -207,7 +207,7 @@ class TestFetchEsiStatus(NoSocketsTestCase):
                 }
             },
             endpoints=[EsiEndpoint("Status", "get_status")],
-            http_error=True,
+            http_error=502,
         )
         # when
         status = fetch_esi_status()